% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nrc_act.R
\name{create_nrc_act}
\alias{create_nrc_act}
\title{create_nrc_act}
\usage{
create_nrc_act()
}
\value{
returns a data frame of school and/or district ACT results.
}
\description{
create_nrc_act creates a data frame with all ACT results and includes state_id, org_id, and
   type. Those fields are not included in the ACT api output and are included by combining it
   with the organization api data.
}
\details{
All of the data, through 2017, that can be pulled by this function is already available in
   a data frame `nrc_act` that comes with the nrc package; View(nrc_act). This function was
   used to create that data frame.
}
