% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nrc_orgs.R
\name{create_nrc_orgs}
\alias{create_nrc_orgs}
\title{create_nrc_orgs}
\usage{
create_nrc_orgs()
}
\value{
returns a data frame of all NRC organizations: schools, dirstricts, and the state.
}
\description{
create_nrc_orgs creates a data frame with all the orgnaizations from the Nevada Report Card (NRC)
   API. This provides an organization id, state id, name, and type for each school, district,
   and the state.
}
\details{
This data is then joined with the demographic data in order to pull in the levels the school
   services (i.e. 1 2 3 for ES, MS, and HS)

All of the data that can be pulled by this function is already available in
   a data frame `nrc_orgs` that comes with the nrc package; View(nrc_orgs). This function was
   used to create that data frame.
}
