% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nrc_sbac.R
\name{create_nrc_sbac}
\alias{create_nrc_sbac}
\title{create_nrc_sbac}
\usage{
create_nrc_sbac()
}
\value{
returns a data frame of school and/or district SBAC results.
}
\description{
create_nrc_sbac creates a data frame with all sbac results and includes state_id, org_id, and
   type. Those fields are not included in the sbac api output and are included by combining it
   with the organization api.
}
\details{
All of the data, through 2017, that can be pulled by this function is already available in
   a data frame `nrc_sbac` that comes with the nrc package; View(nrc_sbac). This function was
   used to create that data frame.
}
