% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_act.R
\name{get_act}
\alias{get_act}
\title{get_act}
\usage{
get_act(org_ids, spring_years = 2016:2018)
}
\arguments{
\item{org_ids}{A numeric vector of NRC organization ids. You can look them up for a school or
district by using get_org_id(name). These can be viewed by looking at the
nrc_orgs data frame that comes with the nrc package.}

\item{spring_years}{a numeric vector of spring school years.
For example, 2016 would be submitted for the 2015-2016 school year and
2016:2017 would be provided to get both the 2015-2016 and 2016-2017 school years.}
}
\value{
returns a data frame of school, district and/or state ACT results.
}
\description{
get_act retrieves ACT results from the Nevada Report Card (NRC) API
}
\details{
WARNING: There are no ids provided with the data extract and there are instances of school
   names that match across school districts.

All of the data, through 2018, that can be pulled by this function is already available in
   a data frame `nrc_act` that comes with the nrc package; View(nrc_act). This function along
   with `create_nrc_act` was used to create that data frame. That data includes state_id,
   org_id, and type.
}
\examples{
# Get district ACT results.
get_act(nrc_orgs$id[nrc_orgs$type == 'D'], spring_years = 2016)
}
