% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collection_id.R
\name{get_collection_id}
\alias{get_collection_id}
\title{get_collection_id}
\usage{
get_collection_id(org_ids)
}
\arguments{
\item{org_ids}{A numeric vector of NRC organization ids. You can look them up for a school or
district by using get_org_id(name). These can be viewed by looking at the
nrc_orgs data frame that comes with the nrc package, View(nrc_orgs).}
}
\value{
returns a collection id for the set of organization ids provided.
}
\description{
get_collection_id returns the Nevada Report Card (NRC) collection id that is used to reference a
   collection of schools and/or districts when requesting data (e.g. assessment data, personnel data, etc.).
   This function is referenced internally by most of the other functions in the nrc package.
}
\examples{
# Get the collection id for all organizations (c18172)
get_collection_id(unique(nrc_orgs$id))

# Get the collection id for organizations in Clark County
get_collection_id(nrc_orgs$id[nrc_orgs$parent_id == 64827])

# Get the collection id for all the districts in Nevadda
get_collection_id(nrc_orgs$id[nrc_orgs$type == 'D'])
}
