% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_demographics.R
\name{get_demographics}
\alias{get_demographics}
\title{get_demographics}
\usage{
get_demographics(org_ids, spring_years = 2004:2017)
}
\arguments{
\item{org_ids}{A numeric vector of NRC organization ids. You can look them up for a school or
district by using get_org_id(name). These can be viewed by looking at the
nrc_orgs data frame that comes with the nrc package.}

\item{spring_years}{a numeric vector of spring school years.
For example, 2016 would be submitted for the 2015-2016 school year and
2015:2016 would be provided to get both the 2014-2015 and 2015-2016 school years.}
}
\value{
returns a data frame of accountability demographics.
}
\description{
get_demographics returns a data frame of accountability demographics from the
   Nevada Report Card (NRC) API given a numeric vector of NRC organization ids.
}
\details{
All of the data that can be pulled by this function is already available in a data frame
   `nrc_demographics` that comes with the nrc package; View(nrc_demographics). This function was
   used to create that data frame.
}
\examples{
all_nde_demographics <- get_demographics(nrc_orgs$id)
}
