% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fiscal.R
\name{get_fiscal}
\alias{get_fiscal}
\title{get_fiscal}
\usage{
get_fiscal(org_ids, spring_years = 2008:2017)
}
\arguments{
\item{org_ids}{A numeric vector of NRC organization ids. You can look them up for a school or
district by using get_org_id(name). These can be viewed by looking at the
nrc_orgs data frame that comes with the nrc package.}

\item{spring_years}{a numeric vector of spring school years.
For example, 2016 would be submitted for the 2015-2016 school year and
2015:2016 would be provided to get both the 2014-2015 and 2015-2016 school years.}
}
\value{
returns a data frame of accountability fiscal data.
}
\description{
get_fiscal returns a data frame of accountability fiscal data from the
   Nevada Report Card (NRC) API given a numeric vector of NRC organization ids.
}
\details{
WARNING! According to NRC, this data is reported for the year prior to the accountability year.
   A column has been added to the API data to indicate report year which is the spring year
   prior to the accountability year.

The data starts in 2007-2008 because footnote on NRC stated that is when district totals began
   to exclude district sponsored charter school data.

All of the data that can be pulled by this function is already available in a data frame
   `nrc_fiscal` that comes with the nrc package; View(nrc_fiscal). This function was
   used to create that data frame.
}
\examples{
all_nde_fiscal <- get_fiscal(nrc_orgs$id)
}
