% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_id.R
\name{get_org_id}
\alias{get_org_id}
\title{get_org_id}
\usage{
get_org_id(name)
}
\arguments{
\item{org_name}{a string that is part of an organization's name (district or school).}
}
\value{
returns a data frame of ids and names for all the organizations that matched
        the org_name that was provided and prints them to the console.
}
\description{
get_org_id returns the Nevada Report Card (NRC) organization ids and names for all organizations
    whose name matches the string provided as the org_name parameter. The organization id is the
    unique id used by NRC, it is different than the state id used by the Nevada Department
    of Education.
}
\details{
The nrc package comes with a data frame `nrc_orgs` that contains all the districts and
    schools that hava a org_id. That is the data set this function pulls information from.  You
    can inspect it by executing: `View(nrc_orgs)`
}
\examples{
# Note the function ignores case (capitalization doesn't matter)
get_org_id('forbuss')
get_org_id('Forbuss')

get_org_id('churchill')
}
