% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_personnel.R
\name{get_personnel}
\alias{get_personnel}
\title{get_personnel}
\usage{
get_personnel(org_ids, spring_years = 2017)
}
\arguments{
\item{org_ids}{A numeric vector of NRC organization ids. You can look them up for a school or
district by using get_org_id(name). These can be viewed by looking at the
nrc_orgs data frame that comes with the nrc package.}

\item{spring_years}{a numeric vector of spring school years.
For example, 2016 would be submitted for the 2015-2016 school year and
2015:2016 would be provided to get both the 2014-2015 and 2015-2016 school years.}
}
\value{
returns a data frame of accountability personnel data.
}
\description{
get_personnel returns a data frame of accountability personnel data from the
   Nevada Report Card (NRC) API given a numeric vector of NRC organization ids and
   spring school years. The function gets teacher ada, sub teacher counts and staffing counts.
}
\details{
All of the data that can be pulled by this function is already available in a data frame
   `nrc_personnel` that comes with the nrc package; View(nrc_personnel). This function was
   used to create that data frame.
}
\examples{
# Get personnel data for CCSD schools for 2015-2016
ccsd_ids <- nrc_orgs$id[nrc_orgs$parent_id == 64827 | nrc_orgs$id == 64827]
ccsd_personnel_2016 <- get_personnel(ccsd_ids, 2016)

# Get personnel data for the state for all years.
state_personnel <- get_personnel(64825, 2004:2016)
}
