% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profile.R
\name{get_profile}
\alias{get_profile}
\title{get_profile}
\usage{
get_profile(org_id, spring_year = 2018)
}
\arguments{
\item{org_id}{A single NRC organization id. You can look them up for a school or
district by using get_org_id(name). These can be viewed by looking at the
nrc_orgs data frame that comes with the nrc package.}

\item{spring_year}{a single spring school year.
For example, 2016 would be submitted for the 2015-2016 school year}
}
\value{
a dataframe of the accountability profile data for the provided entity and school year.
}
\description{
get_profile returns the school profile from the Nevada Report Card (NRC) API given the school's
    organization id and a spring school year.
}
\examples{
# Get the 2013-2014 profile for Forbuss ES
forbuss_org_id <- get_org_id('forbuss')$org_id
forbuss_profile <- get_profile(org_id = forbuss_org_id, spring_year = 2014)
}
