% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sbac.R
\name{get_sbac}
\alias{get_sbac}
\title{get_sbac}
\usage{
get_sbac(org_ids, spring_years = 2016:2018, grades = 3:8)
}
\arguments{
\item{org_ids}{A numeric vector of NRC organization ids. You can look them up for a school or
district by using get_org_id(name). These can be viewed by looking at the
nrc_orgs data frame that comes with the nrc package.}

\item{spring_years}{a numeric vector of spring school years.
For example, 2016 would be submitted for the 2015-2016 school year and
2015:2016 would be provided to get both the 2014-2015 and 2015-2016 school years.}

\item{grades}{a numeric vector of grades to be included. Can be numbers 3-8}
}
\value{
returns a data frame of school and/or district SBAC results.
}
\description{
get_sbac retrieves SBAC results from the Nevada Report Card (NRC) API
}
\details{
WARNING: There are no ids provided with the data extract and there are instances of school
   names that match across school districts.

All of the data, through 2018, that can be pulled by this function is already available in
   a data frame `nrc_sbac` that comes with the nrc package; View(nrc_sbac). This function along
   with `create_nrc_sbac` was used to create that data frame. That data includes state_id,
   org_id, and type.
}
\examples{
# Get district SBAC results for all grades.
get_sbac(nrc_orgs$id[nrc_orgs$type == 'D'], spring_years = 2016, grades = 3:8)
}
