% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sbac_subgroup.R
\name{get_sbac_subgroup}
\alias{get_sbac_subgroup}
\title{get_sbac_subgroup}
\usage{
get_sbac_subgroup(org_ids, spring_years = 2016:2017, grades = 3:8,
  ethnicities = NULL, gender = NULL, iep = NULL, frl = NULL,
  ell = NULL, condition = "and")
}
\arguments{
\item{org_ids}{A numeric vector of NRC organization ids. You can look them up for a school or
district by using get_org_id(name). These can be viewed by looking at the
nrc_orgs data frame that comes with the nrc package.}

\item{spring_years}{a numeric vector of spring school years.
For example, 2016 would be submitted for the 2015-2016 school year and
2015:2016 would be provided to get both the 2014-2015 and 2015-2016 school years.}

\item{grades}{a numeric vector of grades to be included. Can be numbers 3-8}

\item{ethnicities}{a character vector of ethnicities to filter by
Possible Values: 'A', 'B', 'C', 'H', 'I', 'M', 'P'
Corresponding to: Asian, Black, Caucasian, Hispanic, Indian (Native American), Multiracial,
Pacific Islander
NULL will not filter by ethnicity.}

\item{gender}{a character of either 'M' or 'F' to filter by male or female.
NULL will not filter by gender.}

\item{iep}{a character of either 'Y' or 'N' to filter by IEP or Not IEP.
NULL will not filter by IEP.}

\item{frl}{a character of either 'Y' or 'N' to filter by FRL or Not FRL.
NULL will not filter by FRL.}

\item{ell}{a character of either 'Y' or 'N' to filter by ELL or Not ELL.
NULL will not filter by ELL.}

\item{condition}{string of either 'AND' or 'OR' that sets the condition on how the other
parameters are combined. (e.g. FRL AND ELL or FRL OR ELL)}
}
\value{
returns a data frame of sbac results or the requested organizations,
   and grades for the filtered subgroup.
}
\description{
get_sbac_subgroup retrieves SBAC subgroup results from the Nevada Report Card (NRC) API
}
\examples{
# Get district results for African American male students.
district_ids <- nrc_orgs$id[nrc_orgs$type == 'D']
african_american_male <- get_sbac_subgroup(org_ids = district_ids, ethnicities = 'b',gender = 'm')
View(african_american_male)

# Get state results for students who are FRL and ELL
state_id <- nrc_orgs$id[nrc_orgs$type == 'S']
state_frl_ell <- get_sbac_subgroup(org_ids = state_id, frl = 'Y', ell = 'Y')
}
