% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_student_data.R
\name{get_student_data}
\alias{get_student_data}
\title{get_student_data}
\usage{
get_student_data(org_ids, spring_years = 2018)
}
\arguments{
\item{org_ids}{A numeric vector of NRC organization ids. You can look them up for a school or
district by using get_org_id(name). These can be viewed by looking at the
nrc_orgs data frame that comes with the nrc package.}

\item{spring_years}{a numeric vector of spring school years.
For example, 2016 would be submitted for the 2015-2016 school year and
2015:2016 would be provided to get both the 2014-2015 and 2015-2016 school years.}
}
\value{
returns a data frame of accountability data.
}
\description{
get_student_data returns a data frame of student accountability data (ADA, Avg class size,
   student/teacher ratio, retention rates, credit deficient, transiency, etc.) from the
   Nevada Report Card (NRC) API given a numeric vector of NRC organization ids and
   spring school years.
}
\examples{
# Get student data for CCSD schools for 2015-2016
ccsd_ids <- nrc_orgs$id[nrc_orgs$parent_id == 64827 | nrc_orgs$id == 64827]
ccsd_student_2016 <- get_student_data(ccsd_ids, 2016)

# Get student data for the state for all years.
state_student <- get_student_data(64825, 2004:2016)
}
