% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_org_labels.R
\name{update_org_labels}
\alias{update_org_labels}
\title{update_org_labels}
\usage{
update_org_labels(org_id, label_name)
}
\value{
Updates the nrc_org data set provided with the `nrc` package.
}
\description{
create_org_labels creats a copy of the nrc_orgs dataset in the global environment named
    `temp_nrc_orgs`.  It then updates the label_name for an organization given that
    organizations id.
}
\details{
You can update a label name and then assign the temp data set back to `nrc_orgs` and
    then use the command `devtools::use_data(nrc_orgs, overwrite == TRUE)` to overwrite the
    set that is embedded in the `nrc` package.  Then `CTRL + SHIFT + B` to rebuild the
    package with the updated nrc_orgs data.
}
